Move groundtruthRegexes module under lib/ and update imports
